A default handler is added and the implementation in the Controller is removed
